Separation of routing and modules
