+++ bot/admin/web/src/app/configuration/configuration.module.ts
@@ -17,13 +17,8 @@
 import { ApplicationsModule } from '../applications/applications.module';
 import { Injectable, NgModule } from '@angular/core';
 import { SharedModule } from '../shared-nlp/shared.module';
-import { RouterModule, Routes } from '@angular/router';
-import { ApplicationsComponent } from '../applications/applications/applications.component';
-import { ApplicationComponent } from '../applications/application/application.component';
-import { AuthGuard } from '../core-nlp/auth/auth.guard';
 import { BotConfigurationsComponent } from './bot/bot-configurations.component';
 import { ConfigurationTabsComponent } from './configuration-tabs.component';
-import { ApplicationsResolver } from '../applications/applications.resolver';
 import { CommonModule } from '@angular/common';
 import { BotSharedModule } from '../shared/bot-shared.module';
 import { BotConfigurationComponent } from './bot/bot-configuration.component';
@@ -44,83 +39,10 @@ import {
 } from '@nebular/theme';
 import { NewBotComponent } from './bot/new-bot.component';
 import { ReactiveFormsModule } from '@angular/forms';
-import { UserLogsComponent } from '../applications/user/user-logs.component';
-import { NamespacesComponent } from '../applications/namespace/namespaces.component';
 import { BotSharedService } from '../shared/bot-shared.service';
 import { ApplicationConfig } from '../applications/application.config';
 import { SelectBotConfigurationDialogComponent } from './bot/selection-dialog/select-bot-configuration-dialog.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    canActivate: [AuthGuard],
-    component: ConfigurationTabsComponent,
-    resolve: {
-      application: ApplicationsResolver
-    },
-    children: [
-      {
-        path: '',
-        component: ApplicationsComponent,
-        resolve: {
-          application: ApplicationsResolver
-        }
-      },
-      {
-        path: 'bot',
-        component: BotConfigurationsComponent,
-        resolve: {
-          application: ApplicationsResolver
-        }
-      },
-      {
-        path: 'create',
-        component: ApplicationComponent
-      },
-      {
-        path: 'nlp',
-        children: [
-          {
-            path: '',
-            component: ApplicationsComponent
-          },
-          {
-            path: 'edit/:id',
-            component: ApplicationComponent
-          },
-          {
-            path: 'create',
-            component: ApplicationComponent
-          }
-        ],
-        resolve: {
-          application: ApplicationsResolver
-        }
-      },
-      {
-        path: 'new',
-        component: NewBotComponent,
-        resolve: {
-          application: ApplicationsResolver
-        }
-      },
-      {
-        path: 'users/logs',
-        component: UserLogsComponent
-      },
-      {
-        path: 'namespaces',
-        component: NamespacesComponent
-      }
-    ]
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class BotConfigurationRoutingModule {}
+import { BotConfigurationRoutingModule } from './configuration-routing.module';
 
 @Injectable()
 export class BotApplicationConfig implements ApplicationConfig {
