+++ nlp/admin/web/src/app/applications/applications.module.ts
@@ -17,8 +17,6 @@
 import { SharedModule } from '../shared-nlp/shared.module';
 import { CommonModule } from '@angular/common';
 import { Injectable, NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { AuthGuard } from '../core-nlp/auth/auth.guard';
 import { ApplicationsComponent } from './applications/applications.component';
 import { ApplicationsResolver } from './applications.resolver';
 import { ApplicationComponent } from './application/application.component';
@@ -46,64 +44,7 @@ import { NamespacesComponent } from './namespace/namespaces.component';
 import { NgJsonEditorModule } from 'ang-jsoneditor';
 import { ApplicationConfig } from './application.config';
 import { NgbPaginationModule } from '@ng-bootstrap/ng-bootstrap';
-
-const routes: Routes = [
-  {
-    path: '',
-    canActivate: [AuthGuard],
-    component: ConfigurationTabsComponent,
-    resolve: {
-      applications: ApplicationsResolver
-    },
-    children: [
-      {
-        path: '',
-        component: ApplicationsComponent,
-        resolve: {
-          application: ApplicationsResolver
-        }
-      },
-      {
-        path: 'create',
-        component: ApplicationComponent
-      },
-      {
-        path: 'nlu',
-        children: [
-          {
-            path: '',
-            component: ApplicationsComponent
-          },
-          {
-            path: 'edit/:id',
-            component: ApplicationComponent
-          },
-          {
-            path: 'create',
-            component: ApplicationComponent
-          }
-        ],
-        resolve: {
-          application: ApplicationsResolver
-        }
-      },
-      {
-        path: 'users/logs',
-        component: UserLogsComponent
-      },
-      {
-        path: 'namespaces',
-        component: NamespacesComponent
-      }
-    ]
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class ApplicationsRoutingModule {}
+import { ApplicationsRoutingModule } from './applications-routing.module';
 
 @Injectable()
 export class NlpApplicationConfig implements ApplicationConfig {
@@ -114,44 +55,44 @@ export class NlpApplicationConfig implements ApplicationConfig {
 }
 
 @NgModule({
-    imports: [
-        CommonModule,
-        SharedModule,
-        MomentModule,
-        ApplicationsRoutingModule,
-        NbTabsetModule,
-        NbCardModule,
-        NbRouteTabsetModule,
-        FileUploadModule,
-        NbCardModule,
-        NbActionsModule,
-        NbButtonModule,
-        NbTooltipModule,
-        NbCheckboxModule,
-        NbSelectModule,
-        NbAccordionModule,
-        NbSpinnerModule,
-        NbRadioModule,
-        NgJsonEditorModule,
-        NbDialogModule.forRoot(),
-        NgbPaginationModule
-    ],
-    declarations: [
-        ApplicationsComponent,
-        ApplicationComponent,
-        ApplicationAdvancedOptionsComponent,
-        ApplicationUploadComponent,
-        UserLogsComponent,
-        DisplayUserDataComponent,
-        ConfigurationTabsComponent,
-        NamespacesComponent
-    ],
-    providers: [
-        {
-            provide: ApplicationConfig,
-            useClass: NlpApplicationConfig
-        },
-        ApplicationsResolver
-    ]
+  imports: [
+    CommonModule,
+    SharedModule,
+    MomentModule,
+    ApplicationsRoutingModule,
+    NbTabsetModule,
+    NbCardModule,
+    NbRouteTabsetModule,
+    FileUploadModule,
+    NbCardModule,
+    NbActionsModule,
+    NbButtonModule,
+    NbTooltipModule,
+    NbCheckboxModule,
+    NbSelectModule,
+    NbAccordionModule,
+    NbSpinnerModule,
+    NbRadioModule,
+    NgJsonEditorModule,
+    NbDialogModule.forRoot(),
+    NgbPaginationModule
+  ],
+  declarations: [
+    ApplicationsComponent,
+    ApplicationComponent,
+    ApplicationAdvancedOptionsComponent,
+    ApplicationUploadComponent,
+    UserLogsComponent,
+    DisplayUserDataComponent,
+    ConfigurationTabsComponent,
+    NamespacesComponent
+  ],
+  providers: [
+    {
+      provide: ApplicationConfig,
+      useClass: NlpApplicationConfig
+    },
+    ApplicationsResolver
+  ]
 })
 export class ApplicationsModule {}
