+++ nlp/admin/web/src/app/core-nlp/auth/auth-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { LoginComponent } from './login/login.component';
+
+const authRoutes: Routes = [
+  {
+    path: 'login',
+    component: LoginComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(authRoutes)],
+  exports: [RouterModule]
+})
+export class AuthRoutingModule {}
