+++ nlp/admin/web/src/app/core-nlp/auth/auth.module.ts
@@ -19,23 +19,10 @@ import { NgModule } from '@angular/core';
 import { LoginComponent } from './login/login.component';
 import { AuthService } from './auth.service';
 import { AuthGuard } from './auth.guard';
-import { RouterModule, Routes } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { RestModule } from '../rest/rest.module';
 import { NbCardModule, NbCheckboxModule } from '@nebular/theme';
-
-const authRoutes: Routes = [
-  {
-    path: 'login',
-    component: LoginComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(authRoutes)],
-  exports: [RouterModule]
-})
-export class AuthRoutingModule {}
+import { AuthRoutingModule } from './auth-routing.module';
 
 @NgModule({
   imports: [CommonModule, SharedModule, RestModule, AuthRoutingModule, NbCardModule, NbCheckboxModule],
